Fixes for qic

+++ cm-mlops/script/get-preprocessed-dataset-imagenet/customize.py
@@ -18,7 +18,7 @@ def preprocess(i):
             print("Using MLCommons Inference source from '" + env['CM_MLPERF_INFERENCE_SOURCE'] +"'")
 
         env['CM_DATASET_PREPROCESSED_PATH'] = os.getcwd()
-        if not exists(os.path.join(env['CM_DATASET_PATH'], "val_map.txt")):
+        if env['CM_DATASET_TYPE'] == "validation" and not exists(os.path.join(env['CM_DATASET_PATH'], "val_map.txt")):
             shutil.copy(os.path.join(env['CM_DATASET_AUX_PATH'], "val.txt"), os.path.join(env['CM_DATASET_PATH'],
             "val_map.txt"))
